fix(web): osk rotation, prevention of page-bottom occlusion
