--- web/src/app/web/README.md
@@ -1,3 +1,0 @@
-**NOTE**: _deprecated_
-
-This subproject holds old namespaced-code corresponding to the new, modularized `app/browser` subproject.
