--- web/src/app/web/tsconfig.json
@@ -1,27 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-
-	"compilerOptions": {
-		"outFile": "../../../build/app/web/obj/keymanweb.js",
-    "sourceRoot": "/keyman/"
-  },
-
-	"include": [
-    "./*.ts",
-    "./**/*.ts"
-  ],
-
-  "files": [
-		"kmwnative.ts"
-  ],
-
-  "references": [
-    { "path": "../../../../common/web/keyman-version", "prepend": true },
-    { "path": "../../../../common/web/utils", "prepend": true },
-    { "path": "../../../../common/predictive-text/browser.tsconfig.json", "prepend": true },
-    { "path": "../../../../common/web/input-processor/src", "prepend": true },
-    { "path": "../../../../common/web/keyboard-processor/src", "prepend": true },
-		{ "path": "../../engine/device-detect", "prepend": true},
-		{ "path": "../../engine/main", "prepend": true}
-	]
-}
